Tooltip: restrict `placement` prop to only "top" & "bottom" - Also inline the TooltipArrow source into the Tooltip.tsx module
