+++ packages/syntax-core/src/Tooltip/Tooltip.module.css
@@ -1,3 +1,12 @@
 .tooltip {
   max-width: 240px;
 }
+
+.arrowPlacementtop {
+  transform: rotate(180deg) translate(-100%, -100%);
+  transform-origin: top left;
+}
+
+.arrowPlacementbottom {
+  transform: rotate(0deg) translate(0, 0);
+}
