--- packages/syntax-core/src/react-aria-utils/AriaOverlayArrow.module.css
@@ -1,8 +1,0 @@
-.placementtop {
-  transform: rotate(180deg) translate(-100%, -100%);
-  transform-origin: top left;
-}
-
-.placementbottom {
-  transform: rotate(0deg) translate(0, 0);
-}
